// frontend/src/api/generated/models/sourceUpdateRequestRefKind.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Palmimo Portal
 * OpenAPI spec version: 0.1.0
 */

export type SourceUpdateRequestRefKind = typeof SourceUpdateRequestRefKind[keyof typeof SourceUpdateRequestRefKind];


export const SourceUpdateRequestRefKind = {
  branch: 'branch',
  tag: 'tag',
} as const;